requirements.txt
# Web Interface
gradio>=4.0.0

# LLM and ML Libraries
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0

# Apple Silicon optimized inference
mlx>=0.0.9
mlx-lm>=0.0.9

# Vector Database
chromadb>=0.4.18

# OpenAI API
openai>=1.3.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
